removed unnecessary files, coordinates are now of double type, typos corrected

+++ app/src/main/java/com/izyasosha/logics/GameObject.java
@@ -8,7 +8,7 @@ import android.graphics.Bitmap;
 
 public abstract class GameObject
 {
-    protected int x=0, y=0;
+    protected double x=0, y=0;
     protected byte size=0;
     protected Bitmap bmp=null;
